requirements.txt
openai==1.5.0
langchain==0.0.350
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.1
streamlit==1.27.2